backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
selenium>=4.25.0
webdriver-manager>=4.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.3.0
